basically v1.0 now ready!
I say basically because there's just one small thing not working, which
is changing coulours to each square, right now if there's more than one
face, all the rectangles are the same colour regardless of the
prediction of each face.
But the project is intended to work for one user anyway, so one could
say v1.0 is basically ready!